pyproject.toml
[tool.poetry]
name = "treeview-django"
version = "0.1.0"
description = ""
authors = ["saalaus <saalaus2@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-dotenv = "^1.4.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.ruff]
line-length = 79
lint.select = [
    "DJ",
]
